refactor: fix name of file and import refs
